resolved PR comments and added rubocop file
