+++ .github/workflows/test.yml
@@ -18,7 +18,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        ruby: [ '2.6', '3.0' ]
+        ruby: [ '2.6', '2.7', '3.0' ]
         os: [ 'macos-latest', 'ubuntu-latest', 'windows-latest']
         experimental: [false]
         include:
