+++ .rubocop.yml
@@ -1,0 +1,2 @@
+inherit_from:
+  - https://raw.githubusercontent.com/riboseinc/oss-guides/master/ci/rubocop.yml
