+++ lib/plurimath.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
-require_relative 'plurimath/version'
-require_relative 'plurimath/plurimath'
+require_relative "plurimath/version"
+require_relative "plurimath/math"
