+++ lib/plurimath/asciimath.rb
@@ -1,6 +1,11 @@
 # frozen_string_literal: true
 
-require_relative 'asciimath_parser'
+require_relative "asciimath/function/sin"
+require_relative "asciimath/function/tan"
+require_relative "asciimath/function/cos"
+require_relative "asciimath/number"
+require_relative "asciimath/symbol"
+require_relative "asciimath/variable"
 module Plurimath
   # Asciimath Class
   class Asciimath
