+++ lib/plurimath/asciimath/function/tan.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Plurimath
+  class Asciimath
+    module Function
+      # Tan Class
+      class Tan
+        attr_accessor :angle
+
+        def initialize(angle)
+          @angle = angle
+        end
+      end
+    end
+  end
+end
