--- lib/plurimath/asciimath_parser.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'asciimath_parser/sin'
-require_relative 'asciimath_parser/tan'
-require_relative 'asciimath_parser/cos'
-require_relative 'asciimath_parser/number'
-require_relative 'asciimath_parser/symbol'
-require_relative 'asciimath_parser/variable'
