--- lib/plurimath/asciimath_parser/cos.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module Plurimath
-  module AsciimathParser
-    # Cos Class
-    class Cos
-      attr_accessor :angle
-
-      def initialize(angle)
-        @angle = angle
-      end
-    end
-  end
-end
