--- lib/plurimath/asciimath_parser/sin.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module Plurimath
-  module AsciimathParser
-    # Sin Class
-    class Sin
-      attr_accessor :angle
-
-      def initialize(angle)
-        @angle = angle
-      end
-    end
-  end
-end
