--- lib/plurimath/asciimath_parser/tan.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module Plurimath
-  module AsciimathParser
-    # Tan Class
-    class Tan
-      attr_accessor :angle
-
-      def initialize(angle)
-        @angle = angle
-      end
-    end
-  end
-end
