--- lib/plurimath/formula.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module Plurimath
-  # Formula Class
-  class Formula
-    attr_accessor :text
-
-    def initialize(text)
-      @text = text
-    end
-  end
-end
