+++ lib/plurimath/math/formula.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Plurimath
+  module Math
+    # Formula Class
+    class Formula
+      attr_accessor :text
+
+      def initialize(text)
+        @text = text
+      end
+    end
+  end
+end
