+++ lib/plurimath/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Plurimath
   VERSION = "0.1.0"
 end
