--- spec/plurimath/plurimath_spec.rb
@@ -1,8 +1,0 @@
-require_relative '../../lib/plurimath/plurimath'
-
-RSpec.describe Plurimath do
-
-  it "raises error on wrong type" do
-    expect{Plurimath.parse("asdf", type: 'wrong_type')}.to raise_error(Plurimath::Error)
-  end
-end
